requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
Django==4.2.2
django-otp==1.2.1
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
multidict==6.0.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
twilio==8.2.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
zope.interface==6.0
